fix: circular deps in help
