+++ src/config/config.ts
@@ -13,7 +13,7 @@ import {Performance} from '../performance'
 import PluginLoader from './plugin-loader'
 import WSL from 'is-wsl'
 import {format} from 'node:util'
-import {getHelpFlagAdditions} from '../help'
+import {getHelpFlagAdditions} from '../help/util'
 import {loadWithData} from '../module-loader'
 import {settings} from '../settings'
 import {stdout} from '../cli-ux/stream'
